pyproject.toml
[project]
name = "ankrag"
version = "0.1.0"
description = "AnkReg: historical RAG for invoice coding on GCP (BigQuery, Gemini, Vertex embeddings / vector search)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform>=1.70.0",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-storage>=2.18.0",
    "google-genai>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "typer>=0.15.0",
]

[project.scripts]
ankrag = "ankrag.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ankrag"]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]
